+++ src/data/THECATAPI_KEY.js
@@ -1,2 +1,8 @@
-// https://thecatapi.com/
+/* API Documentations
+ * https://thecatapi.com/
+ * https://developers.thecatapi.com/view-account/ylX4blBYT9FaoVd6OhvR?report=bOoHBz-8t
+ * https://documenter.getpostman.com/view/5578104/RWgqUxxh
+ * Key needed for specific parameters
+ * For example, 'order=DESC' or 'order=ASC'
+ */
 export default "live_yR409wwiyuO78r7I8KniHPeLZgVlucehXCttabgRVRlAGluyrwpph9hl4LOsQTB2";